refactor: TestUtil 관련 리팩토링
